Move functions into new Python modules. Integrate with DynamicalDataset
